REFACTOR: split schemas into the folders
