+++ controllers/users/index.js
@@ -1,10 +1,5 @@
-const register = require("./register");
-
 const UsersController = require("./UsersController")
 
-const { ctrlWrapper } = require("../../helpers")
-
 module.exports = {
-  // register: ctrlWrapper(register),
   UsersController
 };
