+++ routes/api/auth.js
@@ -5,20 +5,20 @@ const { isValidId, validateBody } = require("../../middlewares");
 // const { usersCtrl } = require("../../controllers")
 const { UsersController} = require("../../controllers");
 
-const {userJoiSchemas} = require("../../schemas")
+const {userRegisterSchema, userLoginSchema} = require("../../schemas/joi")
 
 
 const authRouter = express.Router();
 
 authRouter.post(
   "/register",
-  validateBody(userJoiSchemas.userRegisterSchema),
+  validateBody(userRegisterSchema),
   UsersController.register
 );
 
 authRouter.post(
   "/login",
-  validateBody(userJoiSchemas.userLoginSchema),
+  validateBody(userLoginSchema),
   UsersController.login
 );
 module.exports = authRouter
