+++ routes/api/contacts.js
@@ -2,7 +2,7 @@ const express = require('express')
 
 const { isValidId, validateBody } = require("../../middlewares");
 const { contactsCtrl } = require("../../controllers");
-const { contactsSchemas } = require("../../schemas");
+const { addSchema, updateFavoriteSchema } = require("../../schemas/joi");
 
 const contactsRouter = express.Router()
 
@@ -12,7 +12,7 @@ contactsRouter.get("/:id", isValidId, contactsCtrl.getById);
 
 contactsRouter.post(
   "/",
-  validateBody(contactsSchemas.addSchema),
+  validateBody(addSchema),
   contactsCtrl.add
 );
 
@@ -21,14 +21,14 @@ contactsRouter.delete("/:id", isValidId, contactsCtrl.removeById);
 contactsRouter.put(
   "/:id",
   isValidId,
-  validateBody(contactsSchemas.addSchema),
+  validateBody(addSchema),
   contactsCtrl.updateById
 );
 
 contactsRouter.patch(
   "/:id",
   isValidId,
-  validateBody(contactsSchemas.updateFavoriteSchema),
+  validateBody(updateFavoriteSchema),
   contactsCtrl.updateStatusContact
 );
