--- schemas/index.js
@@ -1,9 +1,0 @@
-const { contactSchema, contactsSchemas } = require("./contacts");
-const { userSchema, userJoiSchemas } = require("./users")
-
-module.exports = {
-  contactSchema,
-  contactsSchemas,
-  userSchema,
-  userJoiSchemas
-};
