+++ schemas/joi/contacts/index.js
@@ -1,0 +1,8 @@
+const addSchema = require("./addSchema");
+
+const updateFavoriteSchema = require("./updateFavoriteSchema");
+
+module.exports = {
+  addSchema,
+  updateFavoriteSchema,
+};
