+++ schemas/joi/index.js
@@ -1,0 +1,10 @@
+const { userLoginSchema, userRegisterSchema } = require("./users");
+
+const {addSchema, updateFavoriteSchema } = require("./contacts")
+
+module.exports = {
+  userLoginSchema,
+  userRegisterSchema,
+  addSchema,
+  updateFavoriteSchema
+};
