+++ schemas/joi/users/index.js
@@ -1,0 +1,7 @@
+const userRegisterSchema = require("./userRegisterSchema");
+const userLoginSchema = require("./userLoginSchema");
+
+module.exports = {
+  userRegisterSchema,
+  userLoginSchema,
+};
