+++ schemas/joi/users/userRegisterSchema.js
@@ -1,0 +1,10 @@
+const Joi = require("joi");
+
+const userRegisterSchema = Joi.object({
+  password: Joi.string().required(),
+  email: Joi.string().required(),
+  subscription: Joi.string(),
+});
+
+
+module.exports = userRegisterSchema
