+++ schemas/mongoose/contacts/contacts.js
@@ -1,0 +1,34 @@
+const { Schema } = require("mongoose");
+
+const { handleMongooseError } = require("../../../helpers");
+
+const contactSchema = new Schema(
+  {
+    name: {
+      type: String,
+      required: [true, "Set name for contact"],
+    },
+    email: {
+      type: String,
+    },
+    phone: {
+      type: String,
+    },
+    favorite: {
+      type: Boolean,
+      default: false,
+    },
+    owner: {
+      type: Schema.Types.ObjectId,
+      ref: "users",
+    },
+  },
+  { versionKey: false, timestamps: true }
+);
+
+contactSchema.post("save", handleMongooseError);
+
+
+
+module.exports = contactSchema 
+
