+++ schemas/mongoose/contacts/index.js
@@ -1,0 +1,3 @@
+const contactSchema = require("./contacts");
+
+module.exports = { contactSchema };
