+++ schemas/mongoose/index.js
@@ -1,0 +1,7 @@
+const { contactSchema } = require("./contacts");
+const { userSchema } = require("./users");
+
+module.exports = {
+  contactSchema,
+  userSchema,
+};
