+++ schemas/mongoose/users/index.js
@@ -1,0 +1,5 @@
+const userSchema = require("./users");
+
+module.exports = {
+  userSchema,
+};
